requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
langchain==0.0.347
openai==0.28.0
python-dotenv==1.0.0
langchain_google_genai
logging